Move the plans adapter to separate class file
